Using size_t, DendroIntL, or unsigned long long where appropriate, instead of unsigned int
This is supposed to fix a crash that was occuring with subDA construction.

+++ FEM/examples/src/heatMat.cpp
@@ -121,7 +121,7 @@ template <unsigned int dim>
 bool HeatMat<dim>::preMatVec(const VECType* in,VECType* out,double scale)
 {
     // apply boundary conditions.
-    std::vector<unsigned int> bdyIndex;
+    std::vector<size_t> bdyIndex;
     m_uiOctDA->getBoundaryNodeIndices(bdyIndex);
 
     for(unsigned int i=0;i<bdyIndex.size();i++)
@@ -132,7 +132,7 @@ template <unsigned int dim>
 bool HeatMat<dim>::postMatVec(const VECType* in,VECType* out,double scale) {
 
     // apply boundary conditions.
-    std::vector<unsigned int> bdyIndex;
+    std::vector<size_t> bdyIndex;
     m_uiOctDA->getBoundaryNodeIndices(bdyIndex);
 
     for(unsigned int i=0;i<bdyIndex.size();i++)
